refactor: 불필요한 클래스 제거
로그인유저, 회원가입유저 클래스를 삭제하고 해당 로직을 함수로 분리
