--- src/ts/vendingMachine/RegisterUser.ts
@@ -1,53 +1,0 @@
-import { REGISTER_ERROR_MESSAGE } from '../constant/errorMessage';
-import { REGISTER } from '../constant/rule';
-import { postRegisterServer } from '../utils/fetchServer';
-
-export type UserInfoType = {
-  email: string;
-  name: string;
-  password: string;
-  confirmPassword: string;
-};
-
-class RegisterUser {
-  async register(userInfo: UserInfoType) {
-    try {
-      return postRegisterServer(userInfo);
-    } catch (error) {
-      throw new Error(REGISTER_ERROR_MESSAGE.FAIL);
-    }
-  }
-
-  validateRegisterBehavior({ email, name, password, confirmPassword }: UserInfoType) {
-    if (this.isBlankInput(email, name, password, confirmPassword)) {
-      throw new Error(REGISTER_ERROR_MESSAGE.BLANK_NOT_ALLOWED);
-    }
-
-    if (this.isExceedNameRange(name)) {
-      throw new Error(REGISTER_ERROR_MESSAGE.EXCEED_NAME_RANGE);
-    }
-
-    if (this.isViolatePasswordRule(password)) {
-      throw new Error(REGISTER_ERROR_MESSAGE.VIOLATE_PASSWORD_RULE);
-    }
-
-    if (this.isNotMatchWithPassword(password, confirmPassword)) {
-      throw new Error(REGISTER_ERROR_MESSAGE.NO_MATCH_PASSWORD_CONFIRM);
-    }
-  }
-
-  private isBlankInput(email: string, name: string, password: string, confirmPassword: string) {
-    return !email.length || !name.length || !password.length || !confirmPassword.length;
-  }
-  private isExceedNameRange(name: string) {
-    return name.length < REGISTER.NAME_MIN_LENGTH || name.length > REGISTER.NAME_MAX_LENGTH;
-  }
-  private isViolatePasswordRule(password: string) {
-    return !REGISTER.PASSWORD_RULE.test(password);
-  }
-  private isNotMatchWithPassword(password: string, confirmPassword: string) {
-    return password !== confirmPassword;
-  }
-}
-
-export default RegisterUser;
